+++ src/Pim/Bundle/InstallerBundle/DataFixtures/AbstractLoadFixturesData.php
@@ -3,12 +3,13 @@
 namespace Pim\Bundle\InstallerBundle\DataFixtures;
 
 use Akeneo\Bundle\BatchBundle\Command\BatchCommand;
-use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
+use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
 use Akeneo\Bundle\BatchBundle\Job\ExitStatus;
 use Doctrine\Common\DataFixtures\AbstractFixture;
 use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
+use Pim\Bundle\InstallerBundle\Exception\JobExecutionException;
 use Pim\Bundle\InstallerBundle\FixtureLoader\FixtureJobLoader;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Input\ArrayInput;
@@ -92,22 +93,39 @@ abstract class AbstractLoadFixturesData extends AbstractFixture implements
                 [
                     'command'     => 'akeneo:batch:job',
                     'code'        => $job->getCode(),
-                    '--no-debug'  => true
+                    '--no-debug'  => true,
+                    '-v'          => true
                 ]
             ),
             new ConsoleOutput()
         );
 
         $execution = $this->getJobExecution($job);
-        if (ExitStatus::COMPLETED !== $execution->getExitStatus()->getExitCode()) {
-            throw new \Exception(
-                sprintf(
-                    'An error occured during execution of the job "%s", detailled trace can be found in %s',
-                    $job->getCode(),
-                    $execution->getLogFile()
-                )
-            );
+        if (!$this->executionComplete($execution)) {
+            throw new JobExecutionException($execution);
         }
+    }
+
+    /**
+     * Returns true if the job has been executed without any warning
+     *
+     * @param JobExecution $jobExecution
+     *
+     * @return true
+     */
+    protected function executionComplete(JobExecution $jobExecution)
+    {
+        if (ExitStatus::COMPLETED !== $jobExecution->getExitStatus()->getExitCode()) {
+            return false;
+        }
+
+        foreach ($jobExecution->getStepExecutions() as $stepExecution) {
+            if (count($stepExecution->getWarnings())) {
+                return false;
+            }
+        }
+
+        return true;
     }
 
     /**